improve: mongodb supports all types of _id field (#505)
* improve: mongodb supports all types of _id field

* test

* fix

* comment

+++ dt-common/src/config/extractor_config.rs
@@ -111,7 +111,7 @@ pub enum ExtractorConfig {
         db_tbs: HashMap<String, Vec<String>>,
         parallel_size: usize,
         parallel_type: RdbParallelType,
-        batch_size: usize,
+        batch_size: u32,
     },
 
     MongoCdc {
